scanner beam activating on key press
