+++ Assets/ScannerBeam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86599c256b3467e4d976af2e712cd17d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
