+++ Assets/Scenes/Scanner Tool Test.unity
@@ -1062,7 +1062,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1388107262}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -0.525, y: 0.283, z: 1.048}
+  m_LocalPosition: {x: -0.525, y: 1.299, z: 1.048}
   m_LocalScale: {x: 0.1, y: 0.1, z: 0.1}
   m_Children:
   - {fileID: 1975212526}
@@ -1081,8 +1081,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 920c4951b88debd4e90fbefcf0c10763, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  materials:
-  - {fileID: 2100000, guid: a7c50eec5064ec74597899cc729adf74, type: 2}
+  beam: {fileID: 0}
 --- !u!54 &1388107268
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -1393,9 +1392,9 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 919132148725006416}
-  m_Enabled: 0
+  m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 920c4951b88debd4e90fbefcf0c10763, type: 3}
+  m_Script: {fileID: 11500000, guid: 86599c256b3467e4d976af2e712cd17d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   materials:
